Dashboard Report Summary & Kanban of Applications

+++ app/Http/Repositories/UserRepository.php
@@ -4,6 +4,7 @@ namespace App\Http\Repositories;
 use App\Http\Contracts\RepositoryContract;
 use App\Models\User;
 use Exception;
+use Illuminate\Support\Facades\Gate;
 
 class UserRepository implements RepositoryContract
 {
@@ -13,6 +14,10 @@ class UserRepository implements RepositoryContract
             $offset    = request()->input('offset') ?? 10;
             $relations = [];
             $fields    = ['id', 'name', 'email', 'phone', 'role', 'status'];
+
+            if (! Gate::allows('view-any', User::class)) {
+                return messageResponse('You are not authorized to view this resource', 403);
+            }
 
             if (request()->has('fields') && request()->input('fields')) {
                 $fields = gettype(request()->input('fields')) === 'array' ? request()->input('fields') : explode(',', request()->input('fields'));
@@ -60,6 +65,10 @@ class UserRepository implements RepositoryContract
                 return messageResponse();
             }
 
+            if (! Gate::allows('view', $doc)) {
+                return messageResponse('You are not authorized to view this resource', 403);
+            }
+
             return entityResponse($doc);
         } catch (Exception $e) {
             return serverError($e);
@@ -69,6 +78,10 @@ class UserRepository implements RepositoryContract
     public function store($request)
     {
         try {
+            if (! Gate::allows('create', User::class)) {
+                return messageResponse('You are not authorized to view this resource', 403);
+            }
+
             User::query()->create($request->validated());
             return messageResponse('User added successfully', 201, 'success');
         } catch (Exception $e) {
@@ -83,6 +96,10 @@ class UserRepository implements RepositoryContract
                 return messageResponse();
             }
 
+            if (! Gate::allows('update', $doc)) {
+                return messageResponse('You are not authorized to view this resource', 403);
+            }
+
             $doc->update($request->validated());
             return messageResponse('User updated successfully', 200, 'success');
         } catch (Exception $e) {
@@ -95,6 +112,10 @@ class UserRepository implements RepositoryContract
         try {
             if (! $doc = User::query()->where(['id' => $id])->first()) {
                 return messageResponse();
+            }
+
+            if (! Gate::allows('delete', $doc)) {
+                return messageResponse('You are not authorized to view this resource', 403);
             }
 
             $doc->delete();
